Move the rest of the definition out of __init__.py.
__init__.py is not intended to contain substantive code. It is intended
declare this a package and to limit the interfaces available to avoid
namespace pollution. It is now doing that.
